Fixes for re-enabling more MSVC level 4 warnings: gpu/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Possibly-uninitialized local variable

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/375733004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281900 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: efc8771d73aae60b93964bfe35a288f0871e210e

+++ gpu/command_buffer/client/transfer_buffer_unittest.cc
@@ -185,7 +185,7 @@ TEST_F(TransferBufferTest, MemoryAlignmentAfterZeroAllocation) {
   Initialize(32u);
   void* ptr = transfer_buffer_->Alloc(0);
   EXPECT_EQ((reinterpret_cast<uintptr_t>(ptr) & (kAlignment - 1)), 0u);
-  transfer_buffer_->FreePendingToken(ptr, -1);
+  transfer_buffer_->FreePendingToken(ptr, static_cast<unsigned int>(-1));
   // Check that the pointer is aligned on the following allocation.
   ptr = transfer_buffer_->Alloc(4);
   EXPECT_EQ((reinterpret_cast<uintptr_t>(ptr) & (kAlignment - 1)), 0u);
